third_party/blink/web_tests/fast/forms/label/contenteditable-label-focus-expected.txt

PASS successfullyParsed is true

TEST COMPLETE

Calling target.focus()
focus received by [object HTMLLabelElement]
PASS document.activeElement is target
I'm editable. Just a button
